Bump min Firefox version to 53 for browserslist
Due to some kagnax compat tests not passing, preset-env considers
Firefox 52 as not capable of supporting generators and async functions.
This results in its adding the dependency on the regeneratorRuntime.
See https://github.com/babel/babel/issues/8204 for details.

The compat tests in question pass in Firefox 53, so the simple fix is to
bump the min supported version to 53. This is a bit unfortunate because
52 was an ESR, but it's also almost 2 years old and there's been ESR 60
for almost half a year now, too.